+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - pin github action to hash instead of explicit version
 - update to yolo11 object detection model
 - fix `opensearch-tunnel` user data install script
+- fix `opensearch-tunnel` names on stack to be dynamic
 - fix `analysis/rosbag-image-pipeline-sfn` to properly record metadata
 - add tunnel and ddb trigger to `rosbag-image-pipeline-sfn`